requirements.txt
langchain 
openai 
pandas
# tabulate
python-dotenv
# openpyxl
streamlit
# pandasai
pypdf
tiktoken
# sentence-transformers
chromadb